Handle Storybook 5.3.x+ Config.
From v5.3.0, Storybook has changed how it implements configuration.

The primary impact is that `config.js` no longer exists; Stories are exported
in `main.js` and the runner is configured from `preview.js`.

This change checks the Storybook version and, if greater then 5.3, will use
`preview.js` to get the Storybook object.  If this file doesn't exist, it will
be created, then removed at the end of the process.